Jet list + sync G5 64-bit
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4690 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ViewerController.h
@@ -424,7 +424,7 @@ enum
 - (IBAction) roiDeleteAll:(id) sender;
 
 /**  methods to access global variables */
-+ (long) numberOf2DViewer;
++ (int) numberOf2DViewer;
 
 // UNDOCUMENTED FUNCTIONS
 // For more informations: rossetantoine@bluewin.ch
